+++ unlicense/application.py
@@ -21,7 +21,7 @@ def main() -> None:
 
 
 def run_unlicense(
-    exe_to_dump: str,
+    pe_to_dump: str,
     verbose: bool = False,
     pause_on_oep: bool = False,
     force_oep: Optional[int] = None,
@@ -29,7 +29,7 @@ def run_unlicense(
     timeout: int = 10,
 ) -> None:
     """
-    Unpack executables protected with WinLicense/Themida 2.x and 3.x
+    Unpack executables protected with Themida/WinLicense 2.x and 3.x
     """
     if verbose:
         log_level = logging.DEBUG
@@ -38,14 +38,14 @@ def run_unlicense(
     logging.basicConfig(level=log_level)
     lief.logging.disable()
 
-    exe_path = Path(exe_to_dump)
-    if not exe_path.is_file():
-        LOG.error("'%s' isn't a file or doesn't exist", exe_path)
+    pe_path = Path(pe_to_dump)
+    if not pe_path.is_file():
+        LOG.error("'%s' isn't a file or doesn't exist", pe_path)
         sys.exit(1)
 
     # Detect Themida/Winlicense version if needed
     if target_version is None:
-        target_version = detect_winlicense_version(exe_to_dump)
+        target_version = detect_winlicense_version(pe_to_dump)
         if target_version is None:
             LOG.error("Failed to automatically detect packer version")
             sys.exit(2)
@@ -55,7 +55,7 @@ def run_unlicense(
     LOG.info("Detected packer version: %d.x", target_version)
 
     # Check PE architecture and bitness
-    if not interpreter_can_dump_pe(exe_to_dump):
+    if not interpreter_can_dump_pe(pe_to_dump):
         LOG.error("Target PE cannot be dumped with this interpreter. "
                   "This is most likely a 32 vs 64 bit mismatch.")
         sys.exit(3)
@@ -76,7 +76,7 @@ def run_unlicense(
 
     # Spawn the packed executable and instrument it to find its OEP
     process_controller = frida_exec.spawn_and_instrument(
-        exe_path, notify_oep_reached)
+        pe_path, notify_oep_reached)
     try:
         # Block until OEP is reached
         if not oep_reached.wait(float(timeout)):
@@ -99,10 +99,10 @@ def run_unlicense(
                 LOG.error(".NET assembly dump failed")
         # Fix imports and dump the executable
         elif target_version == 2:
-            winlicense2.fix_and_dump_pe(process_controller, exe_to_dump,
+            winlicense2.fix_and_dump_pe(process_controller, pe_to_dump,
                                         dumped_image_base, dumped_oep)
         elif target_version == 3:
-            winlicense3.fix_and_dump_pe(process_controller, exe_to_dump,
+            winlicense3.fix_and_dump_pe(process_controller, pe_to_dump,
                                         dumped_image_base, dumped_oep)
     finally:
         # Try to kill the process on exit