+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- Handle unpacking of 32-bit and 64-bit DLLs
 - Handle unpacking of 32-bit and 64-bit .NET assembly PEs (EXE only)
 - OEP detection times out after 10 seconds by default. The duration can be
   changed through the CLI.