+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.2.0] - 2022-05-31
 ### Added
 - Handle unpacking of 32-bit and 64-bit DLLs
 - Handle unpacking of 32-bit and 64-bit .NET assembly PEs (EXE only)